move api/processes/steps/ssj.js into ssj subfolder
- also stubbed a function to show dashboard example
